refactor: move drive factory
