+++ common/drive_handle.go
@@ -1,9 +1,8 @@
-package go_dparm
+package common
 
 import (
 	"errors"
 	"github.com/jc-lab/go-dparm/ata"
-	"github.com/jc-lab/go-dparm/common"
 	"github.com/jc-lab/go-dparm/internal"
 	"github.com/jc-lab/go-dparm/nvme"
 	"github.com/jc-lab/go-dparm/tcg"
@@ -15,17 +14,17 @@ import (
 const trimSet = " \t\r\n\x00"
 
 type DriveHandleImpl struct {
-	common.DriveHandle
-	dh   common.DriverHandle
-	Info common.DriveInfo
+	DriveHandle
+	Dh   DriverHandle
+	Info DriveInfo
 }
 
 var (
 	ErrNotSupportThisDriver = errors.New("not supported in this driver")
 )
 
-func (p *DriveHandleImpl) init() error {
-	ataDrive, ok := p.dh.(common.AtaDriverHandle)
+func (p *DriveHandleImpl) Init() error {
+	ataDrive, ok := p.Dh.(AtaDriverHandle)
 	if ok {
 		identityRaw := ataDrive.GetIdentity()
 		p.Info.AtaIdentityRaw = identityRaw
@@ -52,7 +51,7 @@ func (p *DriveHandleImpl) init() error {
 		p.Info.IsSsd = p.Info.SsdCheckWeight > 0
 	}
 
-	nvmeDrive, ok := p.dh.(common.NvmeDriverHandle)
+	nvmeDrive, ok := p.Dh.(NvmeDriverHandle)
 	if ok {
 		identityRaw := nvmeDrive.GetIdentity()
 		p.Info.NvmeIdentityRaw = identityRaw
@@ -77,14 +76,14 @@ func (p *DriveHandleImpl) init() error {
 	return nil
 }
 
-func (p *DriveHandleImpl) GetDriverHandle() common.DriverHandle {
-	return p.dh
+func (p *DriveHandleImpl) GetDriverHandle() DriverHandle {
+	return p.Dh
 }
 
 func (p *DriveHandleImpl) Close() {
-	if p.dh != nil {
-		p.dh.Close()
-		p.dh = nil
+	if p.Dh != nil {
+		p.Dh.Close()
+		p.Dh = nil
 	}
 }
 
@@ -92,20 +91,20 @@ func (p *DriveHandleImpl) GetDevicePath() string {
 	return p.Info.DevicePath
 }
 
-func (p *DriveHandleImpl) GetDrivingType() common.DrivingType {
-	return p.dh.GetDrivingType()
+func (p *DriveHandleImpl) GetDrivingType() DrivingType {
+	return p.Dh.GetDrivingType()
 }
 
 func (p *DriveHandleImpl) GetDriverName() string {
-	return p.dh.GetDriverName()
+	return p.Dh.GetDriverName()
 }
 
-func (p *DriveHandleImpl) GetDriveInfo() *common.DriveInfo {
+func (p *DriveHandleImpl) GetDriveInfo() *DriveInfo {
 	return &p.Info
 }
 
 func (p *DriveHandleImpl) AtaDoTaskFileCmd(rw bool, dma bool, tf *ata.Tf, data []byte, timeoutSecs int) error {
-	impl, ok := p.dh.(common.AtaDriverHandle)
+	impl, ok := p.Dh.(AtaDriverHandle)
 	if ok {
 		return impl.DoTaskFileCmd(rw, dma, tf, data, timeoutSecs)
 	}
@@ -113,7 +112,7 @@ func (p *DriveHandleImpl) AtaDoTaskFileCmd(rw bool, dma bool, tf *ata.Tf, data [
 }
 
 func (p *DriveHandleImpl) NvmeGetLogPage(nsid uint32, logId uint32, rae bool, size int) ([]byte, error) {
-	impl, ok := p.dh.(common.NvmeDriverHandle)
+	impl, ok := p.Dh.(NvmeDriverHandle)
 	if ok {
 		return impl.NvmeGetLogPage(nsid, logId, rae, size)
 	}
@@ -121,16 +120,16 @@ func (p *DriveHandleImpl) NvmeGetLogPage(nsid uint32, logId uint32, rae bool, si
 }
 
 func (p *DriveHandleImpl) SecurityCommand(rw bool, dma bool, protocol uint8, comId uint16, buffer []byte, timeoutSecs int) error {
-	if p.dh == nil {
+	if p.Dh == nil {
 		return errors.New("Not supported")
 	}
 
-	err := p.dh.SecurityCommand(rw, dma, protocol, comId, buffer, timeoutSecs)
+	err := p.Dh.SecurityCommand(rw, dma, protocol, comId, buffer, timeoutSecs)
 	if err == nil {
 		return nil
 	}
 
-	ataDriver, ok := p.dh.(common.AtaDriverHandle)
+	ataDriver, ok := p.Dh.(AtaDriverHandle)
 	if ok {
 		tf := &ata.Tf{}
 		tf.Lob.Feat = protocol
@@ -146,7 +145,7 @@ func (p *DriveHandleImpl) SecurityCommand(rw bool, dma bool, protocol uint8, com
 		return ataDriver.DoTaskFileCmd(rw, dma, tf, buffer, timeoutSecs)
 	}
 
-	nvmeDriver, ok := p.dh.(common.NvmeDriverHandle)
+	nvmeDriver, ok := p.Dh.(NvmeDriverHandle)
 	if ok {
 		cmd := &nvme.NvmeAdminCmd{}
 		cmd.Opcode = uint8(internal.Ternary(rw, nvme.NVME_ADMIN_OP_SECURITY_SEND, nvme.NVME_ADMIN_OP_SECURITY_RECV))
